add relay xcm configs
