+++ relay/polkadot/src/xcm_config.rs
@@ -49,7 +49,6 @@ use xcm_builder::{
 pub use pallet_rc_migrator::xcm_config::{AssetHubLocation, CollectivesLocation};
 
 parameter_types! {
-	pub const RootLocation: Location = Here.into_location();
 	/// The location of the DOT token, from the context of this chain. Since this token is native to this
 	/// chain, we make it synonymous with it and thus it is the `Here` location, which means "equivalent to
 	/// the context".
@@ -155,13 +154,6 @@ impl Contains<Location> for OnlyParachains {
 	}
 }
 
-pub struct LocalPlurality;
-impl Contains<Location> for LocalPlurality {
-	fn contains(loc: &Location) -> bool {
-		matches!(loc.unpack(), (0, [Plurality { .. }]))
-	}
-}
-
 /// The barriers one of which must be passed for an XCM message to be executed.
 pub type Barrier = TrailingSetTopicAsId<(
 	// Weight that is paid for may be consumed.
@@ -181,10 +173,6 @@ pub type Barrier = TrailingSetTopicAsId<(
 		ConstU32<8>,
 	>,
 )>;
-
-/// Locations that will not be charged fees in the executor, neither for execution nor delivery.
-/// We only waive fees for system functions, which these locations represent.
-pub type WaivedLocations = (SystemParachains, Equals<RootLocation>, LocalPlurality);
 
 pub struct XcmConfig;
 impl xcm_executor::Config for XcmConfig {
@@ -215,7 +203,9 @@ impl xcm_executor::Config for XcmConfig {
 	type PalletInstancesInfo = AllPalletsWithSystem;
 	type MaxAssetsIntoHolding = MaxAssetsIntoHolding;
 	type FeeManager = XcmFeeManagerFromComponents<
-		WaivedLocations,
+		pallet_rc_migrator::xcm_config::WaivedLocations<crate::RcMigrator>,
+		// TODO: post-ahm move the Treasury funds from this local account to sovereign account
+		// of the new AH Treasury.
 		SendXcmFeeToAccount<Self::AssetTransactor, TreasuryAccount>,
 	>;
 	// No bridges on the Relay Chain
